+++ src/lib/components/events/UbicationTicketsCard.svelte
@@ -1,5 +1,4 @@
 <script lang="ts">
-  import { Ticket } from "lucide-svelte";
   interface Ticket {
     index: number;
     amount: number;
@@ -22,9 +21,10 @@
     style: "currency",
     currency: "CLP",
   }).format(ticketPrice);
+
 </script>
 
-<div
+<!-- <div
   class:opacity-25={ticket.amount === 0}
   class={`w-full border p-2 indicator flex flex-col justify-center items-center pt-2 max-w-sm min-h-16 ${
     ticket.type === "ringside_tickets" ? "border-success" : "border-info"
@@ -62,5 +62,29 @@
       >
       <p class="text-lg">{formatedPrice || 0}</p>
     {/if}
+  </div>
+</div> -->
+
+<div class:opacity-30={ticket.amount === 0} class="w-full">
+  <div class="relative bg-primary p-6 text-black" >
+    <!-- Notches -->
+    <div class="absolute left-0 top-16 w-4 h-4 bg-black" />
+    <div class="absolute right-0 top-16 w-4 h-4 bg-black" />
+
+    <h2 class="text-2xl font-bold mb-2 text-center pb-2">{ticket.name}</h2>
+    <div class="w-full border border-black/30 border-b-0 border-dashed"></div>
+    <div class="flex justify-center gap-4 items-center mb-4 mt-2">
+      <p class="text-2xl font-bold text-center line-through opacity-50">
+        {dicountPercentage ? `$${priceBeforeDiscount}` : ""}
+      </p>
+      <p class="text-2xl font-bold text-center">{formatedPrice || 0}</p>
+    </div>
+    <p
+      class={`mt-2 font-semibold tracking-widest uppercase text-sm text-center`}
+    >
+      {ticket.amount > 1
+        ? `Últimas ${ticket.amount} entradas`
+        : "Agotado"}
+    </p>
   </div>
 </div>